React with animation
