+++ app/components/customControlBar.tsx
@@ -1,9 +1,7 @@
 import {
   ControlBar,
   ControlBarProps,
-  useLocalParticipant,
   useLocalParticipantPermissions,
-  useRoomContext,
 } from "@livekit/components-react";
 
 import { TrackSource } from "livekit-server-sdk";
@@ -11,26 +9,16 @@ import useIsAdmin from "@/app/hooks/useIsAdmin";
 import { AdminControls } from "./admin/adminControls";
 import { Modal } from "./modal";
 import { PauseButton } from "./admin/adminControls/pauseButton";
-import { Participant, RemoteParticipant, RoomEvent } from "livekit-client";
+import { EmojiHandler } from "./emojiHandler";
 
 interface CustomControlBarProps extends ControlBarProps {
   token: string;
 }
 
 export function CustomControlBar({ token, ...props }: CustomControlBarProps) {
-  const localParticipant = useLocalParticipant();
-  const encoder = new TextEncoder();
-  const decoder = new TextDecoder();
-  const room = useRoomContext();
-
   const permissions = useLocalParticipantPermissions();
   const isAdmin = useIsAdmin(token);
 
-  room.on(RoomEvent.DataReceived, (payload, participant, kind) => {
-    const strData = decoder.decode(payload);
-    // console.log(`Got message from ${participant?.identity}`);
-    console.log(`Message is ${strData}`);
-  });
   return (
     <div className="lk-control-bar">
       {isAdmin && (
@@ -44,17 +32,7 @@ export function CustomControlBar({ token, ...props }: CustomControlBarProps) {
           <PauseButton />
         </>
       )}
-      <button
-      className="lk-button"
-        onClick={() => {
-          localParticipant.localParticipant.publishData(
-            encoder.encode("Hello"),
-            { reliable: true, topic: "EmojiReaction" }
-          );
-        }}
-      >
-        Emoji React
-      </button>
+      <EmojiHandler />
       <ControlBar
         controls={{
           camera: permissions?.canPublishSources.includes(TrackSource.CAMERA),
